+++ sqllin-driver/src/nativeMain/kotlin/com/ctrip/sqllin/driver/cinterop/NativeStatement.kt
@@ -66,28 +66,31 @@ internal class NativeStatement(
 ) : SQLiteStatement {
 
     // Cursor methods
-    fun isNull(index: Int): Boolean =
+    private fun isNull(index: Int): Boolean =
         sqlite3_column_type(cStatementPointer, index) == SQLITE_NULL
 
-    override fun columnGetLong(columnIndex: Int): Long =
-        sqlite3_column_int64(cStatementPointer, columnIndex)
+    override fun columnGetLong(columnIndex: Int): Long? =
+        if (isNull(columnIndex)) null else sqlite3_column_int64(cStatementPointer, columnIndex)
 
-    override fun columnGetDouble(columnIndex: Int): Double =
-        sqlite3_column_double(cStatementPointer, columnIndex)
+    override fun columnGetDouble(columnIndex: Int): Double? =
+        if (isNull(columnIndex)) null else sqlite3_column_double(cStatementPointer, columnIndex)
 
-    override fun columnGetString(columnIndex: Int): String =
-        sqlite3_column_text(cStatementPointer, columnIndex)
-            ?.reinterpret<ByteVar>()
-            ?.let { bytesToString(it) }
-            ?: ""
+    override fun columnGetString(columnIndex: Int): String? =
+        if (isNull(columnIndex))
+            null
+        else
+            sqlite3_column_text(cStatementPointer, columnIndex)
+                ?.reinterpret<ByteVar>()
+                ?.let { bytesToString(it) }
 
-    override fun columnGetBlob(columnIndex: Int): ByteArray {
+    override fun columnGetBlob(columnIndex: Int): ByteArray? {
+        if (isNull(columnIndex))
+            return null
         val blobSize = sqlite3_column_bytes(cStatementPointer, columnIndex)
         return if (blobSize == 0)
-            byteArrayOf()
+            null
         else
             sqlite3_column_blob(cStatementPointer, columnIndex)?.readBytes(blobSize)
-                ?: throw sqliteException("Byte array size/type issue col $columnIndex")
     }
 
     override fun columnCount(): Int = sqlite3_column_count(cStatementPointer)